[bug 1097175] Reorganized search view tests into separate files.
